moved general template functionality to template_engine
